+++ src/main/java/com/ivanov/kirill/EmployeeHandbook/controller/OrganizationController.java
@@ -4,11 +4,13 @@ import com.ivanov.kirill.EmployeeHandbook.dto.OrganizationDto;
 import com.ivanov.kirill.EmployeeHandbook.email.EmailService;
 import com.ivanov.kirill.EmployeeHandbook.model.Organization;
 import com.ivanov.kirill.EmployeeHandbook.service.OrganizationService;
+import com.ivanov.kirill.EmployeeHandbook.utils.AuthorizeUser;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.*;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -63,6 +65,12 @@ public class OrganizationController {
     public ResponseEntity<String> deleteOrganizationById(
             @RequestParam Long id
     ) {
+        Optional<Organization> organization = organizationService.getOrganizationById(id);
+        if (!organization.isPresent())
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Invalid id");
+        if (!AuthorizeUser.checkOrganizationAccess(organization.get(), SecurityContextHolder.getContext().getAuthentication().getPrincipal()))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Not enough permissions");
+
         if (organizationService.deleteOrganization(id)) {
             emailService.sendMailMessage(
                     "kirdmiv@gmail.com",
@@ -97,8 +105,14 @@ public class OrganizationController {
             @PathVariable Long id,
             @RequestBody OrganizationDto organizationRequest
     ) {
-        Organization organization = modelMapper.map(organizationRequest, Organization.class);
-        Optional<Organization> updatedOrganization = organizationService.updateOrganization(id, organization);
+        Optional<Organization> organization = organizationService.getOrganizationById(id);
+        if (!organization.isPresent())
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+        if (!AuthorizeUser.checkOrganizationAccess(organization.get(), SecurityContextHolder.getContext().getAuthentication().getPrincipal()))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+
+        Organization newOrganization = modelMapper.map(organizationRequest, Organization.class);
+        Optional<Organization> updatedOrganization = organizationService.updateOrganization(id, newOrganization);
         if (updatedOrganization.isPresent()) {
             OrganizationDto updatedOrganizationDto = modelMapper.map(updatedOrganization.get(), OrganizationDto.class);
             emailService.sendMailEntity(
